+++ spec/models/SearchParamsModel_spec.js
@@ -275,6 +275,7 @@ define([
         expect(model.get('title')).toBe(defaultParams.osTitle);
         expect(model.get('startDate')).toBe(defaultParams.osDtStart);
         expect(model.get('endDate')).toBe(defaultParams.osDtEnd);
+        expect(model.get('geoBoundingBox')).toBe(defaultParams.osGeoBbox);
       });
 
     });
@@ -336,7 +337,7 @@ define([
       });
 
       it('resets all search parameters to their initial state', function () {
-        model.reset();
+        model.resetCriteria();
 
         expect(model.get('pageNumber')).toEqual(1);
         expect(model.get('keyword')).toBe('');